Refactor the Code.
1. Added few more class library project in the application for Maintainability and Usability of the application.

2. Created the “Org.Feeder.App.Model” library to maintain all model of the “Feeder” application which were added in “Framework” folder of “Org.Feeder.App” application.

3. Created “Org.Feeder.App.Service” library to provide the data to the “FeederApplication”. This is working as factory to get data via using data provider.

This is dependent upon the configuration of app.config file later we can change and provide configuration at run time so that it works independently.

Earlier, Navigator class has the responsibility to get data from the database that I have changed.

4. Created two data provider “Org.Feeder.App.SampeDataProvider” and “Org.Feeder.App.DataProvider”.

Sample Data Provider: created this data provider to showcase to use multiple data provider and it’s use in the “Feeder” application. Also, we can use the same as mocking the data in our Test project if required or for testing application in offline mode if required.

Data Provider: This is the wrapper of existing database dll provided with the application to get data from database. We are assuming that we are only getting time out exception from database side so I haven’t added more on exception side accept database connectivity issue.

5. Along with that used “AutoFac” Ioc Container in the application to remove dependency. I am not able to use much of this because of small application and less dependency in the project.